+++ matcher/src/pattern/tests.rs
@@ -1,20 +1,20 @@
-use crate::pattern::{Atom, AtomKind, CaseMatching};
+use crate::pattern::{Atom, AtomKind, CaseMatching, Normalization};
 
 #[test]
 fn negative() {
-    let pat = Atom::parse("!foo", CaseMatching::Smart);
+    let pat = Atom::parse("!foo", CaseMatching::Smart, Normalization::Smart);
     assert!(pat.negative);
     assert_eq!(pat.kind, AtomKind::Substring);
     assert_eq!(pat.needle.to_string(), "foo");
-    let pat = Atom::parse("!^foo", CaseMatching::Smart);
+    let pat = Atom::parse("!^foo", CaseMatching::Smart, Normalization::Smart);
     assert!(pat.negative);
     assert_eq!(pat.kind, AtomKind::Prefix);
     assert_eq!(pat.needle.to_string(), "foo");
-    let pat = Atom::parse("!foo$", CaseMatching::Smart);
+    let pat = Atom::parse("!foo$", CaseMatching::Smart, Normalization::Smart);
     assert!(pat.negative);
     assert_eq!(pat.kind, AtomKind::Postfix);
     assert_eq!(pat.needle.to_string(), "foo");
-    let pat = Atom::parse("!^foo$", CaseMatching::Smart);
+    let pat = Atom::parse("!^foo$", CaseMatching::Smart, Normalization::Smart);
     assert!(pat.negative);
     assert_eq!(pat.kind, AtomKind::Exact);
     assert_eq!(pat.needle.to_string(), "foo");
@@ -22,23 +22,23 @@ fn negative() {
 
 #[test]
 fn pattern_kinds() {
-    let pat = Atom::parse("foo", CaseMatching::Smart);
+    let pat = Atom::parse("foo", CaseMatching::Smart, Normalization::Smart);
     assert!(!pat.negative);
     assert_eq!(pat.kind, AtomKind::Fuzzy);
     assert_eq!(pat.needle.to_string(), "foo");
-    let pat = Atom::parse("'foo", CaseMatching::Smart);
+    let pat = Atom::parse("'foo", CaseMatching::Smart, Normalization::Smart);
     assert!(!pat.negative);
     assert_eq!(pat.kind, AtomKind::Substring);
     assert_eq!(pat.needle.to_string(), "foo");
-    let pat = Atom::parse("^foo", CaseMatching::Smart);
+    let pat = Atom::parse("^foo", CaseMatching::Smart, Normalization::Smart);
     assert!(!pat.negative);
     assert_eq!(pat.kind, AtomKind::Prefix);
     assert_eq!(pat.needle.to_string(), "foo");
-    let pat = Atom::parse("foo$", CaseMatching::Smart);
+    let pat = Atom::parse("foo$", CaseMatching::Smart, Normalization::Smart);
     assert!(!pat.negative);
     assert_eq!(pat.kind, AtomKind::Postfix);
     assert_eq!(pat.needle.to_string(), "foo");
-    let pat = Atom::parse("^foo$", CaseMatching::Smart);
+    let pat = Atom::parse("^foo$", CaseMatching::Smart, Normalization::Smart);
     assert!(!pat.negative);
     assert_eq!(pat.kind, AtomKind::Exact);
     assert_eq!(pat.needle.to_string(), "foo");
@@ -46,69 +46,69 @@ fn pattern_kinds() {
 
 #[test]
 fn case_matching() {
-    let pat = Atom::parse("foo", CaseMatching::Smart);
+    let pat = Atom::parse("foo", CaseMatching::Smart, Normalization::Smart);
     assert!(pat.ignore_case);
     assert_eq!(pat.needle.to_string(), "foo");
-    let pat = Atom::parse("Foo", CaseMatching::Smart);
+    let pat = Atom::parse("Foo", CaseMatching::Smart, Normalization::Smart);
     assert!(!pat.ignore_case);
     assert_eq!(pat.needle.to_string(), "Foo");
-    let pat = Atom::parse("Foo", CaseMatching::Ignore);
+    let pat = Atom::parse("Foo", CaseMatching::Ignore, Normalization::Smart);
     assert!(pat.ignore_case);
     assert_eq!(pat.needle.to_string(), "foo");
-    let pat = Atom::parse("Foo", CaseMatching::Respect);
+    let pat = Atom::parse("Foo", CaseMatching::Respect, Normalization::Smart);
     assert!(!pat.ignore_case);
     assert_eq!(pat.needle.to_string(), "Foo");
-    let pat = Atom::parse("Foo", CaseMatching::Respect);
+    let pat = Atom::parse("Foo", CaseMatching::Respect, Normalization::Smart);
     assert!(!pat.ignore_case);
     assert_eq!(pat.needle.to_string(), "Foo");
-    let pat = Atom::parse("Äxx", CaseMatching::Ignore);
+    let pat = Atom::parse("Äxx", CaseMatching::Ignore, Normalization::Smart);
     assert!(pat.ignore_case);
     assert_eq!(pat.needle.to_string(), "äxx");
-    let pat = Atom::parse("Äxx", CaseMatching::Respect);
+    let pat = Atom::parse("Äxx", CaseMatching::Respect, Normalization::Smart);
     assert!(!pat.ignore_case);
-    let pat = Atom::parse("Axx", CaseMatching::Smart);
+    let pat = Atom::parse("Axx", CaseMatching::Smart, Normalization::Smart);
     assert!(!pat.ignore_case);
     assert_eq!(pat.needle.to_string(), "Axx");
-    let pat = Atom::parse("你xx", CaseMatching::Smart);
+    let pat = Atom::parse("你xx", CaseMatching::Smart, Normalization::Smart);
     assert!(pat.ignore_case);
     assert_eq!(pat.needle.to_string(), "你xx");
-    let pat = Atom::parse("你xx", CaseMatching::Ignore);
+    let pat = Atom::parse("你xx", CaseMatching::Ignore, Normalization::Smart);
     assert!(pat.ignore_case);
     assert_eq!(pat.needle.to_string(), "你xx");
-    let pat = Atom::parse("Ⲽxx", CaseMatching::Smart);
+    let pat = Atom::parse("Ⲽxx", CaseMatching::Smart, Normalization::Smart);
     assert!(!pat.ignore_case);
     assert_eq!(pat.needle.to_string(), "Ⲽxx");
-    let pat = Atom::parse("Ⲽxx", CaseMatching::Ignore);
+    let pat = Atom::parse("Ⲽxx", CaseMatching::Ignore, Normalization::Smart);
     assert!(pat.ignore_case);
     assert_eq!(pat.needle.to_string(), "ⲽxx");
 }
 
 #[test]
 fn escape() {
-    let pat = Atom::parse("foo\\ bar", CaseMatching::Smart);
+    let pat = Atom::parse("foo\\ bar", CaseMatching::Smart, Normalization::Smart);
     assert_eq!(pat.needle.to_string(), "foo bar");
-    let pat = Atom::parse("\\!foo", CaseMatching::Smart);
+    let pat = Atom::parse("\\!foo", CaseMatching::Smart, Normalization::Smart);
     assert_eq!(pat.needle.to_string(), "!foo");
     assert_eq!(pat.kind, AtomKind::Fuzzy);
-    let pat = Atom::parse("\\'foo", CaseMatching::Smart);
+    let pat = Atom::parse("\\'foo", CaseMatching::Smart, Normalization::Smart);
     assert_eq!(pat.needle.to_string(), "'foo");
     assert_eq!(pat.kind, AtomKind::Fuzzy);
-    let pat = Atom::parse("\\^foo", CaseMatching::Smart);
+    let pat = Atom::parse("\\^foo", CaseMatching::Smart, Normalization::Smart);
     assert_eq!(pat.needle.to_string(), "^foo");
     assert_eq!(pat.kind, AtomKind::Fuzzy);
-    let pat = Atom::parse("foo\\$", CaseMatching::Smart);
+    let pat = Atom::parse("foo\\$", CaseMatching::Smart, Normalization::Smart);
     assert_eq!(pat.needle.to_string(), "foo$");
     assert_eq!(pat.kind, AtomKind::Fuzzy);
-    let pat = Atom::parse("^foo\\$", CaseMatching::Smart);
+    let pat = Atom::parse("^foo\\$", CaseMatching::Smart, Normalization::Smart);
     assert_eq!(pat.needle.to_string(), "foo$");
     assert_eq!(pat.kind, AtomKind::Prefix);
-    let pat = Atom::parse("\\^foo\\$", CaseMatching::Smart);
+    let pat = Atom::parse("\\^foo\\$", CaseMatching::Smart, Normalization::Smart);
     assert_eq!(pat.needle.to_string(), "^foo$");
     assert_eq!(pat.kind, AtomKind::Fuzzy);
-    let pat = Atom::parse("\\!^foo\\$", CaseMatching::Smart);
+    let pat = Atom::parse("\\!^foo\\$", CaseMatching::Smart, Normalization::Smart);
     assert_eq!(pat.needle.to_string(), "!^foo$");
     assert_eq!(pat.kind, AtomKind::Fuzzy);
-    let pat = Atom::parse("!\\^foo\\$", CaseMatching::Smart);
+    let pat = Atom::parse("!\\^foo\\$", CaseMatching::Smart, Normalization::Smart);
     assert_eq!(pat.needle.to_string(), "^foo$");
     assert_eq!(pat.kind, AtomKind::Substring);
 }